fix(playground): make src/server.ts the Vercel Bun-runtime entrypoint
The deployed function 500'd on every route with "Invalid export found in
module". Root cause: Vercel's Bun backend mode auto-detects a root entrypoint
by filename (server/index/app in root or src/), and picked src/server.ts —
which had no default export — over the intended api/index.ts. The build log
said "Using src/server.ts as the root entrypoint", but that module exposed no
valid { fetch } handler, so the runtime found the file and 500'd.

Fix: make src/server.ts itself the entrypoint Vercel already chose.
- Add a default export { fetch: handleRequest } to src/server.ts (the Web
  Standard shape Vercel's Bun runtime calls). The import.meta.main dev/CLI
  block is untouched and never runs on Vercel (imported, not main).
- Point vercel.json functions at src/server.ts (keeping includeFiles for
  packages/components) and drop the unnecessary rewrites table — the Bun root
  entrypoint is a catch-all for every route.
- Remove the redundant api/index.ts indirection; vercel-build now smoke-imports
  src/server.ts; move/rewrite the config test to src/.
- Replace an unsafe `Object.keys(allMaps) as ArtifactFamily[]` cast in
  findArtifactForFamily with entry iteration (the playground lint-staged now
  runs type-aware oxlint, which flags the cast).
- Update CONTRIBUTING.md to document the root-entrypoint model.

Verified: src/server.ts imports side-effect-free (no port bound), default.fetch
delegates /ping -> pong, vercel-build passes, tsc + lint clean, server tests
pass.
